+++ index.html
@@ -1,15 +1,28 @@
 <!DOCTYPE html>
-<html>
-
+<html lang="en">
 <head>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width">
-    <title>Import Export Practice</title>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Inmport Export Pracrice</title>
+    <link rel="stylesheet" href="./style.css">  
 </head>
 
 <body>
+
     <h1 class="shadow">Artists</h1>
-    <div id="list"></div>
+    <div id="list">
+        <ul>
+            <li>
+            <div class="container">
+            <img class="profile 1" src="./images/bob-ross.jpg" alt="bob the painter"> 
+            <img class="profile 2" src="./images/frida-kahlo.jpg" alt="picture of frida kahlo">
+            <img class="profile 3" src="./images/dali.jpg" alt="pictue of dali">
+            </div>
+            </li>
+        </ul>
+    </div>
+    <script type="module" src="./js/app.js" ></script>
+    <script src="./js/personalProfile.js"></script>
 </body>
 
 </html>
