+++ js/personalProfile.js
@@ -1,16 +1,18 @@
-
-class Person {
+export class Person {
     constructor(name, imageUrl) {
         this.name = name;
         this.imageUrl = imageUrl;
     }
-}
 
-function render(person) {
+    modul
+
+  render() {
     return `
         <div class="profile shadow">
-            <img src="${person.imageUrl}"/>
-            <h2>${person.name}</h2>
+            <img src="${this.imageUrl}"/>
+            <h2>${this.name}</h2>
         </div>
+        
     `
 }
+  }
